Merge pull request #196 from OZEO-DOOZ/develop

+++ ios/Classes/DoozMeshManagerApi/Channels/Arguments/SendGenericLevelGetArguments.swift
@@ -6,6 +6,6 @@
 //
 
 struct SendGenericLevelGetArguments: BaseFlutterArguments {
-    let address: Int16
-    let keyIndex: Int16
+    let address: Int
+    let keyIndex: Int
 }
